Fixed major convention issues

+++ wav/wav/Helpers/removeUserIDFromCoreData.swift
@@ -10,17 +10,17 @@ import CoreData
 import UIKit
 
 func removeUserIDFromCoreData() {
-    let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
-    
-    let fetchRequest: NSFetchRequest<User> = User.fetchRequest()
-    
-    do {
-        let users = try context.fetch(fetchRequest)
-        if let user = users.first {
-            context.delete(user)
-            try context.save()
+    if let appDelegate = UIApplication.shared.delegate as? AppDelegate {
+        let context = appDelegate.persistentContainer.viewContext
+        let fetchRequest: NSFetchRequest<User> = User.fetchRequest()
+        do {
+            let users = try context.fetch(fetchRequest)
+            if let user = users.first {
+                context.delete(user)
+                try context.save()
+            }
+        } catch {
+            print("Error removing user ID from Core Data: \(error)")
         }
-    } catch {
-        print("Error removing user ID from Core Data: \(error)")
     }
 }
